+++ react/CheckPermission.tsx
@@ -22,16 +22,19 @@ function CheckPermission({
     ssr: false,
     skip: !roles.length,
   })
-  
+
   if (error) {
     console.error('CheckPermission error:', error)
+
     return null
   }
 
   if (!roles.length || !data || (called && loading)) {
-    return LoadingContent
-      ? <LoadingContent />
-      : <ExtensionPoint id="loading-content" />
+    return LoadingContent ? (
+      <LoadingContent />
+    ) : (
+      <ExtensionPoint id="loading-content" />
+    )
   }
 
   const hasPermission = roles?.length