+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Possible loading content while permissions are loaded.
+
 ### Removed
 - [ENGINEERS-1247] - Disable cypress tests in PR level